requirements.txt
torch
numpy
scipy
matplotlib
librosa
auraloss
pyyaml
tqdm
tensorboard
wheel
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
pyroomacoustics
